requirements.txt
fastapi
uvicorn
yt-dlp
firebase-admin
python-dotenv
requests
implicit
scipy
pandas
numpy
